+++ t/apicast-policy-routing.t
@@ -2305,3 +2305,114 @@ yay, api backend
 --- error_code: 200
 --- no_error_log
 [error]
+
+
+=== TEST 34: Match all conditions working as expected
+Context: https://issues.redhat.com/browse/THREESCALE-6415
+--- configuration
+{
+  "services": [
+    {
+      "id": 42,
+      "proxy": {
+        "policy_chain": [
+          {
+            "name": "apicast.policy.routing",
+            "configuration": {
+              "rules": [
+                {
+                  "url": "http://test:$TEST_NGINX_SERVER_PORT/",
+                  "condition": {
+                    "operations": [
+                      {
+                        "match": "path",
+                        "op": "==",
+                        "value": "/a_path"
+                      }
+                    ]
+                  }
+                },
+
+                {
+                  "url": "http://test:$TEST_NGINX_SERVER_PORT/",
+                  "condition": {}
+                }
+              ]
+            }
+          }
+        ]
+      }
+    }
+  ]
+}
+--- upstream
+  location ~ / {
+     content_by_lua_block {
+       ngx.say(ngx.var.uri)
+     }
+  }
+--- request eval
+["GET /a_path", "GET /"]
+--- response_body eval
+["/a_path\n", "/\n"]
+--- error_code eval
+[200, 200]
+--- no_error_log
+[error]
+
+
+=== TEST 35: Match all conditions working as expected
+Context: https://issues.redhat.com/browse/THREESCALE-6415
+operations is an empty array instead of nil
+--- configuration
+{
+  "services": [
+    {
+      "id": 42,
+      "proxy": {
+        "policy_chain": [
+          {
+            "name": "apicast.policy.routing",
+            "configuration": {
+              "rules": [
+                {
+                  "url": "http://test:$TEST_NGINX_SERVER_PORT/",
+                  "condition": {
+                    "operations": [
+                      {
+                        "match": "path",
+                        "op": "==",
+                        "value": "/a_path"
+                      }
+                    ]
+                  }
+                },
+
+                {
+                  "url": "http://test:$TEST_NGINX_SERVER_PORT/",
+                  "condition": {
+                    "operations": []
+                  }
+                }
+              ]
+            }
+          }
+        ]
+      }
+    }
+  ]
+}
+--- upstream
+  location ~ / {
+     content_by_lua_block {
+       ngx.say(ngx.var.uri)
+     }
+  }
+--- request eval
+["GET /a_path", "GET /"]
+--- response_body eval
+["/a_path\n", "/\n"]
+--- error_code eval
+[200, 200]
+--- no_error_log
+[error]